requirements.txt
scikit-learn~=1.4.0
tensorflow~=2.15.0
numpy~=1.26.4
matplotlib~=3.8.2
pillow~=10.2.0